Extract hook for removing cart item
